+++ src/OnePointCrossover.cpp
@@ -13,9 +13,9 @@ std::unique_ptr<SimpleSchedule> OnePointCrossover<SimpleSchedule>::cross(
     const SimpleSchedule& a, const SimpleSchedule& b) const {
   std::unique_ptr<SimpleSchedule> s = std::make_unique<SimpleSchedule>(a);
 
-  int n = s->size();
-  int idx = Random::randint() % n;
-  for (int i = idx; i < n; i++) {
+  size_t n = s->size();
+  size_t idx = Random::randint() % n;
+  for (size_t i = idx; i < n; i++) {
     s->ires[i] = b.ires[i];
   }
 
@@ -29,9 +29,9 @@ std::unique_ptr<PrioSchedule> OnePointCrossover<PrioSchedule>::cross(
     const PrioSchedule& a, const PrioSchedule& b) const {
   std::unique_ptr<PrioSchedule> s = std::make_unique<PrioSchedule>(a);
 
-  int n = s->size();
-  int idx = Random::randint() % n;
-  for (int i = idx; i < n; i++)
+  size_t n = s->size();
+  size_t idx = Random::randint() % n;
+  for (size_t i = idx; i < n; i++)
     s->ires[i] = b.ires[i];
 
   return std::move(s);
