+++ src/Project.cpp
@@ -7,25 +7,25 @@ namespace SchedulingProblem {
 
 std::unique_ptr<Project> Project::project_ = nullptr;
 
-void Project::create(int _n,
+void Project::create(size_t _n,
                      const std::vector<Task> &_tasks,
-                     int _res_count,
+                     size_t _res_count,
                      const std::vector<int> &_res_id,
                      const std::vector<double> &_res_sal) {
   project_ = std::unique_ptr<Project>(
       new Project(_n, _tasks, _res_count, _res_id, _res_sal));
 
   /* Create list of inverse dependency_. */
-  std::vector<int> *next = new std::vector<int>[_n];
+  std::vector<size_t> *next = new std::vector<size_t>[_n];
 
-  for (int i = 0; i < _n; ++i) {
-    for (int j = 0; j < _tasks[i].num_dependencies(); ++j) {
-      int idep = task(i).dependency(j);
+  for (size_t i = 0; i < _n; ++i) {
+    for (size_t j = 0; j < _tasks[i].num_dependencies(); ++j) {
+      size_t idep = task(i).dependency(j);
       next[idep].push_back(i);
     }
   }
 
-  for (int i = 0; i < _n; ++i) {
+  for (size_t i = 0; i < _n; ++i) {
     task(i).next = next[i];
   }
